refactor: TargetDuration의 함수와 useState 위치 수정
