+++ src/pages/TodayFocusPage/TodayFocus.jsx
@@ -12,6 +12,14 @@ const TodayFocus = () => {
   const [timer, setTimer] = useState(targetDuration);
   const [timerStatus, setTimerStatus] = useState('CANCELED');
   const { id } = useParams();
+  const h = Math.floor((targetDuration / (1000 * 60 * 60)) % 24);
+  const m = Math.floor((targetDuration / (1000 * 60)) % 60);
+  const s = Math.floor((targetDuration / 1000) % 60);
+  const [toggleForm, setToggleForm] = useState('DEFAULT');
+  const [hours, setHours] = useState(h);
+  const [minutes, setMinutes] = useState(m);
+  const [seconds, setSeconds] = useState(s);
+  const [error, setError] = useState('');
 
   const timerRef = useRef();
 
@@ -20,6 +28,9 @@ const TodayFocus = () => {
       const getTimer = await upsertTimer(Number(id));
 
       setTargetDuration(getTimer.targetDuration);
+      if (getTimer.status === 'CANCELED') {
+        setTimer(getTimer.targetDuration);
+      }
     };
 
     fetchTimer();
@@ -41,6 +52,87 @@ const TodayFocus = () => {
     setTimerStatus('IN_PROGRESS');
   };
 
+  // 폼 토글 핸들러 (클릭 시 수정 폼으로 변환)
+  const toggleFormHandler = () => {
+    if (timerStatus !== 'IN_PROGRESS') {
+      if (toggleForm === 'DEFAULT') {
+        setHours(h);
+        setMinutes(m);
+        setSeconds(s);
+        setToggleForm('FORM');
+      } else {
+        setToggleForm('DEFAULT');
+      }
+    }
+  };
+
+  /*---------------------------
+            input 핸들러
+      -------------------------*/
+  const hoursInputHandler = (e) => {
+    const newHours = e.target.value;
+    if (isNaN(newHours)) {
+      setError('숫자를 입력해주세요');
+    } else if (newHours >= 24) {
+      setError('23시 이하로 입력해주세요');
+    } else {
+      setHours(newHours);
+      setError('');
+    }
+  };
+
+  const minutesInputHandler = (e) => {
+    const newMinutes = e.target.value;
+    if (isNaN(newMinutes)) {
+      setError('숫자를 입력해주세요');
+    } else if (newMinutes >= 60) {
+      setError('59분 이하로 입력해주세요');
+    } else {
+      setMinutes(newMinutes);
+      setError('');
+    }
+  };
+
+  const secondsInputHandler = (e) => {
+    const newSeconds = e.target.value;
+    if (isNaN(newSeconds)) {
+      setError('숫자를 입력해주세요');
+    } else if (newSeconds >= 60) {
+      setError('59초 이하로 입력해주세요');
+    } else {
+      setSeconds(newSeconds);
+      setError('');
+    }
+  };
+
+  // 목표 시간 설정 핸들러
+  const submitHandler = async (e) => {
+    e.preventDefault();
+    if (!hours) {
+      setHours(0);
+    }
+    if (!minutes) {
+      setMinutes(0);
+    }
+    if (!seconds) {
+      setSeconds(0);
+    }
+
+    const formattedMs =
+      Number(hours) * 1000 * 60 * 60 +
+      Number(minutes) * 1000 * 60 +
+      Number(seconds) * 1000;
+
+    if (formattedMs < 600000) {
+      setError('10분 이상으로 입력해주세요');
+      return;
+    }
+
+    setTargetDuration(formattedMs);
+    setToggleForm('DEFAULT');
+    setError('');
+  };
+
   return (
     <div className="wrapper">
       <div className={styles.focusWrapper}>
@@ -53,8 +145,17 @@ const TodayFocus = () => {
             <h2>오늘의 집중</h2>
             <TargetDuration
               targetDuration={targetDuration}
-              setTargetDuration={setTargetDuration}
-              timerStatus={timerStatus}
+              toggleForm={toggleForm}
+              error={error}
+              setError={setError}
+              hours={hours}
+              minutes={minutes}
+              seconds={seconds}
+              onToggleForm={toggleFormHandler}
+              onChangeHours={hoursInputHandler}
+              onChangeMinutes={minutesInputHandler}
+              onChangeSeconds={secondsInputHandler}
+              onSubmitTarget={submitHandler}
             />
           </div>
           <Timer
